% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLsingle.R
\name{eQTLsingle}
\alias{eQTLsingle}
\title{eQTLsingle: Discover single-cell eQTLs from scRNA-seq data only}
\usage{
eQTLsingle(
  expressionMatrix,
  snv.gene.pair.metadata,
  p.adjust.method = "bonferroni"
)
}
\arguments{
\item{expressionMatrix}{A dataframe, describes gene expressions (gene expressions by cells), rownames of this dataframe are Geneid, colnames of this dataframe are CellId}

\item{snv.gene.pair.metadata}{the snv-gene pair for testing, can be generated by the function eQTLsingle_build_metadata}

\item{p.adjust.method}{Method for adjusting P-values for multiple comparisons. Default method is bonferroni correction}
}
\value{
A dataframe, each row describes eQTL discovering result of a snv-gene pair
\itemize{
\item SNVid: Id of SNV, represented as CHR__POS
\item Geneid: gene id of target gene
\item sample_size_Ref, sample_size_Alt: number of cells in REF and ALT group
\item theta_Ref, theta_Alt, mu_Ref, mu_Alt, size_Ref, size_Alt, prob_Ref, prob_Alt: estimated parameters of ZINB for REF group and ALT group
\item total_mean_Ref, total_mean_Alt: mean of the REF group and ALT group
\item foldChange: fold change of mean of REF group (total_mean_Ref) with respect to mean of ALT group (total_mean_Alt)
\item chi2LR1: chi square statistic for the test
\item pvalue, adjusted_pvalue: pvalue and adjusted pvalue of the test. If adjusted p-value is smaller than some threshold, this SNV shows significant eQTL effect on the target gene
\item Remark: to record abnormity
}
}
\description{
A function to discover eQTLs from scRNA-seq data. The SNV-gene pair information is generated with function eQTLsingle_build_metadata.
}
\examples{
# Generate metadata for further eQTL analysis.
# Load snvMatrix and expressionMatrix
data(toy_snvMatrix)
data(toy_expressionMatrix)
# Only test the SNVs in which cells with either genotypes (REF and ALT) are at least 30 cells
# Only test the genes which express in > 30 cells in group REF and ALT with expression level >1
snv.gene.pair.metadata <- eQTLsingle_build_metadata(toy_snvMatrix,
                                                    toy_expressionMatrix,
                                                    snv.number.of.cells=30,
                                                    expression.min=1,
                                                    expression.number.of.cells=30)
# Discover eQTLs
eQTL.result <- eQTLsingle(toy_expressionMatrix, snv.gene.pair.metadata)
}
