% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLsingle_build_metadata.R
\name{eQTLsingle_build_metadata}
\alias{eQTLsingle_build_metadata}
\title{Generate metadata for single-cell eQTL analysis.}
\usage{
eQTLsingle_build_metadata(
  snvMatrix,
  expressionMatrix,
  snv.number.of.cells = 30,
  expression.min = 1,
  expression.number.of.cells = 30
)
}
\arguments{
\item{snvMatrix}{A dataframe, describes genotype of SNVs (loci by cells), rownames of this dataframe are SNVid, colnames of this dataframe are Cell ids.}

\item{expressionMatrix}{A dataframe, describes gene expressions (gene expressions by cells), rownames of this dataframe are Geneid, colnames of this dataframe are CellId}

\item{snv.number.of.cells}{threshold of minimal number of cells where mutation (or non-mutation) occurs}

\item{expression.min}{threshold of expression levels of valid genes, used along with another parameter, expression.number.of.cells}

\item{expression.number.of.cells}{threshold of minimal number of cells in which valid genes express.}
}
\value{
A dataframe, each row describes information of valid SNV-gene pairs and corresponding valid cells for eQTL analysis
\itemize{
\item SNVid: Id of SNV, represented as CHR__POS
\item Num_cells_ref: number of cells in REF group
\item Num_cells_alt: number of cells in ALT group
\item Ref_cells: cell list of REF group
\item Alt_cells: cell list of ALT group
\item GeneList: gene list, these genes will be tested with the SNV in this row
\item Num_gene: number of genes in gene list
\item CellList: whole cell list of REF and ALT group
}
}
\description{
This function generates a dataframe describing metadata for further eQTL analysis. The metadata includes valid SNV-gene pairs and corresponding valid cells for eQTL analysis.
}
\examples{
# Generate metadata for further eQTL analysis.
# Load snvMatrix and expressionMatrix
data(toy_snvMatrix)
data(toy_expressionMatrix)
# Only test the SNVs in which cells with either genotypes (REF and ALT) are at least 30 cells
# Only test the genes which express in > 30 cells in group REF and ALT with expression level >1
snv.gene.pair.metadata <- eQTLsingle_build_metadata(toy_snvMatrix,
                                                    toy_expressionMatrix,
                                                    snv.number.of.cells=30,
                                                    expression.min=1,
                                                    expression.number.of.cells=30)
}
