% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLsingle_visualization.R
\name{eQTLsingle_visualization}
\alias{eQTLsingle_visualization}
\title{To visualize eQTL effect}
\usage{
eQTLsingle_visualization(
  eqtl.result,
  expressionMatrix,
  snv.gene.pair.metadata,
  figure.output = NULL
)
}
\arguments{
\item{eqtl.result}{a dataframe (only one row), describes eQTL result of one SNV-gene pair, can be generated from the function eQTLsingle}

\item{expressionMatrix}{dataframe of gene expression}

\item{snv.gene.pair.metadata}{snv-gene pair metadata, generated from the function eQTLsingle_build_metadata}

\item{figure.output}{optional parameter, path of saving the figure}
}
\value{
violin plots of gene expression in both ALT group and REF group
}
\description{
This function can generate violin plots of gene expression in both REF group and ALT group
}
\examples{
# Generate metadata for further eQTL analysis.
# Load snvMatrix and expressionMatrix
data(toy_snvMatrix)
data(toy_expressionMatrix)
# Only test the SNVs in which cells with either genotypes (REF and ALT) are at least 30 cells
# Only test the genes which express in > 30 cells in group REF and ALT with expression level >1
snv.gene.pair.metadata <- eQTLsingle_build_metadata(toy_snvMatrix,
                                                    toy_expressionMatrix,
                                                    snv.number.of.cells=30,
                                                    expression.min=1,
                                                    expression.number.of.cells=30)
# Discover eQTLs
eQTL.result <- eQTLsingle(toy_expressionMatrix, snv.gene.pair.metadata)
# to draw violin plots of SNV (with snvid, "SNV_3") and gene (with geneid, "gene_16")
eQTL.result.target <- eQTL.result[(eQTL.result$SNVid == "SNV_3") &
                                  (eQTL.result$Geneid == "gene_16"),]
eQTLsingle_visualization(eQTL.result.target,
                         toy_expressionMatrix,
                         snv.gene.pair.metadata)
}
