% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_readcounts_tidy.R
\name{snv_readcounts_tidy}
\alias{snv_readcounts_tidy}
\title{Tidy results of eQTLsingle-SNVPostAnalysis.sh}
\usage{
snv_readcounts_tidy(readcounts.result, minimal.read.num = 5)
}
\arguments{
\item{readcounts.result}{The output of eQTLsingle-SNVPostAnalysis.sh}

\item{minimal.read.num}{The threshold of minimal number of reads of each base (A, C, G, T) over the locus}
}
\value{
A data frame, describes number of reads of each base over different loci in single cells
\itemize{
\item CHROM: chromosome
\item POS: position
\item REF: reference base
\item ALT: alternative base
\item DP: depth (number of reads over the locus)
\item ADepth: number of reads with base A over the locus
\item CDepth: number of reads with base C over the locus
\item GDepth: number of reads with base G over the locus
\item TDepth: number of reads with base T over the locus
\item CELL: cell
\item Genotype: a binary variable, describing the genotype of the locus. 'REF' indicates the genotype of the locus is non-mutated (all of reads over it are same as reference). 'ALT' indicates the genotype of this locus is mutated (some of reads over it are same as alternative allele)
}
}
\description{
This function is to tidy output from the script eQTLsingle-SNVPostAnalysis.sh into a data frame.
}
