% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_expressionMatrix}
\alias{toy_expressionMatrix}
\title{expression Matrix of 200 cells}
\format{
A dataset describing expression information for 20 genes across 200 single cells
}
\usage{
data(toy_expressionMatrix)
}
\description{
A test dataset describing expression Matrix of 200 cells.
}
\keyword{data}
