% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_snvMatrix}
\alias{toy_snvMatrix}
\title{snv Matrix of 200 cells}
\format{
A dataset describing mutation information for 10 SNVs across 200 single cells
}
\usage{
data(toy_snvMatrix)
}
\description{
A test dataset describing snv Matrix of 200 cells.
}
\keyword{data}
